# Removed unused **kwargs from GEOSFuncFactory.__call__().