--- a/django/contrib/gis/geos/libgeos.py
+++ b/django/contrib/gis/geos/libgeos.py
@@ -150,8 +150,8 @@ class GEOSFuncFactory:
         self.args = args
         self.kwargs = kwargs
 
-    def __call__(self, *args, **kwargs):
-        return self.func(*args, **kwargs)
+    def __call__(self, *args):
+        return self.func(*args)
 
     @cached_property
     def func(self):