This slightly improves performance:

Before:
```
$ python -m pyperf timeit -s 'from django.contrib.gis.geos import Point; p = Point()' "p.empty"  --duplicate=100
.....................
Mean +- std dev: 2.00 us +- 0.04 us
```

After:
```
$ python -m pyperf timeit -s 'from django.contrib.gis.geos import Point; p = Point()' "p.empty"  --duplicate=100
.....................
Mean +- std dev: 1.94 us +- 0.03 us
```